Sync (#10)
* Fix awarding "There are two of them!" (#1454)

Contrary to what its description stated, this badge was awarded on any
encounter with multiple opponents.

With this commit, it is now only awarded when the dragons succeed after
fledging (not when attacking redundaknights). This is consistent with
the description.

* Fix detecting the '5' key with numpads and azerty leopards

Clicking the beach requires to press 'f' then '5'. The detection of the
'5' key in all possible configurations is a bit tricky, because the
KeyboardEvent interface is such a mess.

Firstly, numpad keys have different keycodes (from 0x60 to 0x69 instead
of 0x30 to 0x39). So we need to check for them individually.

Secondly, on LaPetite's weird azerty leopard, digits in the top row
needs to be typed while holding shift (or control with some layouts).
Control keys are considered normal keys by onkeydown, so the code
remembers them as the previous key: the sequence becomes 'f' shift '5'.
So we test for these keys and ignore them.

* Allow +10/-10 nest buttons to set the value to maximum / minimum

The "double arrow" buttons on Dragon Nest completely refused to work if
they couldn't add or substract 10 percents. They now do the sensible
thing: set the value to its limit (0, 100, or the maximum imposed by
other linings).

* Fixed syntax error in sokoban primitives (missin commas)

* removed some redundancy from redundancy.js
B === D updated to b === D

* new version 4.01

* Set default log length to previous value

* Fix a typo in Topiary's description

* move Changelog from README.MD to its own file

* add Changelog.md

* Fix missing spaces in Polarizer's description

* Fix spelling of the notification when getting a splosion